refactor(packages): correct the package boundaries
Three moves that all fix the same mistake in both directions: agnostic code sitting in the `SqlServer` package, and provider specific code sitting in the agnostic one. Doing them together means consumers change their `using` directives once.

- Entity type configurations split across both packages. Auditing the individual steps, exactly three are SQL Server specific: `IsClustered`, the `NEWID()` default and the `sysname` audit columns.
- Everything else is EF Core or EF Core Relational and works unchanged on other providers.
- The agnostic bases now live in `BB84.EntityFrameworkCore.Repositories.Configurations`; the SqlServer ladder derives from them and adds only those three.
- Type names are identical in both namespaces, so the namespace carries the distinction and existing SQL Server consumers change one `using`.
- Both ladders are complete: every rung exists in both namespaces even where the SqlServer counterpart adds nothing, so that picking a namespace is never a trap.
- The interceptors move to `BB84.EntityFrameworkCore.Repositories` outright. Neither had any SQL Server dependency.
- `DatabaseFacadeExtensions` moves the other way, into the SqlServer package. It emits `EXECUTE [schema].[name] @p = @p OUTPUT` and `SELECT [Value] = [schema].[fn](@p)`, which is T-SQL by construction.
- While moving it, the parameter sequence is now materialized once.
- Previously it was enumerated to build the SQL and enumerated again to bind the command, so a lazily generated sequence produced placeholders that did not match the parameters it was given.
- The duplicate key declaration in the enumerator configuration goes away with the split.
- Clustering is the only job of the SqlServer rungs now, so re-declaring the key is what those rungs are for rather than a redundant second call inside one class.

**Verified behavior preserving:** the create script generated from the test model is byte identical before and after.
