+++ src/BB84.EntityFrameworkCore.Repositories.SqlServer/Configurations/CompositeConfiguration.cs
@@ -1,4 +1,4 @@
-﻿// Copyright: 2024 Robert Peter Meyer
+// Copyright: 2024 Robert Peter Meyer
 // License: MIT
 //
 // This source code is licensed under the MIT license found in the
@@ -8,25 +8,17 @@ using System.Diagnostics.CodeAnalysis;
 using BB84.EntityFrameworkCore.Entities.Abstractions;
 
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+using Base = BB84.EntityFrameworkCore.Repositories.Configurations;
 
 namespace BB84.EntityFrameworkCore.Repositories.SqlServer.Configurations;
 
-/// <summary>
-/// Represents an abstract base class for configuring entity types that implement the
-/// <see cref="ICompositeEntity"/> interface.
-/// </summary>
+/// <inheritdoc cref="Base.CompositeConfiguration{TEntity}"/>
 /// <remarks>
-/// This class is intended to be used as a base for defining entity type configurations in
-/// Entity Framework Core. It provides a default implementation for configuring common properties,
-/// such as the <c>Timestamp</c> property.
+/// Nothing in this configuration is SQL Server specific. The type exists so that the ladder is
+/// complete in both namespaces and a consumer can pick the namespace rather than the type.
 /// </remarks>
-/// <typeparam name="TEntity">The type of the entity being configured.</typeparam>
 [SuppressMessage("Style", "IDE0058", Justification = "Not relevant here, entity type configuration.")]
-public abstract class CompositeConfiguration<TEntity> : IEntityTypeConfiguration<TEntity>
+public abstract class CompositeConfiguration<TEntity> : Base.CompositeConfiguration<TEntity>, IEntityTypeConfiguration<TEntity>
 	where TEntity : class, ICompositeEntity
-{
-	/// <inheritdoc/>
-	public virtual void Configure(EntityTypeBuilder<TEntity> builder)
-		=> EntityTypeBuilderDefaults.ApplyConcurrencyToken(builder, columnOrder: 3);
-}
+{ }
