+++ src/BB84.EntityFrameworkCore.Repositories.SqlServer/Configurations/EntityTypeBuilderDefaults.cs
@@ -14,69 +14,29 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 namespace BB84.EntityFrameworkCore.Repositories.SqlServer.Configurations;
 
 /// <summary>
-/// Provides the shared property configuration used by the entity type configuration base classes.
+/// Provides the SQL Server specific property configuration used by the entity type configuration
+/// base classes of this package.
 /// </summary>
 /// <remarks>
-/// The configuration base classes form ladders of progressively narrower generic aliases. The rungs
-/// share no common base that pins the creator and key types to their defaults, so the defaults that
-/// depend on those types live here instead of being repeated on every rung.
+/// The provider-agnostic counterpart lives in the <c>BB84.EntityFrameworkCore.Repositories</c>
+/// package. Only defaults that cannot be expressed without SQL Server belong here.
 /// </remarks>
 [SuppressMessage("Style", "IDE0058", Justification = "Not relevant here, entity type configuration.")]
 internal static class EntityTypeBuilderDefaults
 {
 	/// <summary>
-	/// Configures the non clustered primary key and the identifier column of the entity.
+	/// Configures whether the primary key of the entity is clustered.
 	/// </summary>
 	/// <typeparam name="TEntity">The type of the entity being configured.</typeparam>
 	/// <typeparam name="TKey">The type of the primary key for the entity.</typeparam>
 	/// <param name="builder">The builder for the entity type being configured.</param>
-	internal static void ApplyIdentityKey<TEntity, TKey>(EntityTypeBuilder<TEntity> builder)
+	/// <param name="clustered">Whether the primary key is clustered.</param>
+	internal static void ApplyKeyClustering<TEntity, TKey>(EntityTypeBuilder<TEntity> builder, bool clustered)
 		where TEntity : class, IIdentity<TKey>
 		where TKey : IEquatable<TKey>
 	{
 		builder.HasKey(e => e.Id)
-			.IsClustered(false);
-
-		builder.Property(e => e.Id)
-			.HasColumnOrder(1)
-			.ValueGeneratedOnAdd();
-	}
-
-	/// <summary>
-	/// Configures the concurrency token of the entity.
-	/// </summary>
-	/// <typeparam name="TEntity">The type of the entity being configured.</typeparam>
-	/// <param name="builder">The builder for the entity type being configured.</param>
-	/// <param name="columnOrder">The zero based ordering of the column within the table.</param>
-	internal static void ApplyConcurrencyToken<TEntity>(EntityTypeBuilder<TEntity> builder, int columnOrder)
-		where TEntity : class, IConcurrency
-	{
-		builder.Property(e => e.Timestamp)
-			.HasColumnOrder(columnOrder)
-			.IsConcurrencyToken()
-			.ValueGeneratedOnAddOrUpdate();
-	}
-
-	/// <summary>
-	/// Configures the creator and editor columns of the entity.
-	/// </summary>
-	/// <typeparam name="TEntity">The type of the entity being configured.</typeparam>
-	/// <typeparam name="TCreator">The type representing the creator of the entity.</typeparam>
-	/// <typeparam name="TEdited">The type representing the editor of the entity.</typeparam>
-	/// <param name="builder">The builder for the entity type being configured.</param>
-	/// <param name="createdByOrder">The ordering of the creator column within the table.</param>
-	/// <param name="editedByOrder">The ordering of the editor column within the table.</param>
-	internal static void ApplyUserAuditColumns<TEntity, TCreator, TEdited>(EntityTypeBuilder<TEntity> builder, int createdByOrder, int editedByOrder)
-		where TEntity : class, IUserAudited<TCreator, TEdited>
-		where TCreator : notnull
-	{
-		builder.Property(e => e.CreatedBy)
-			.HasColumnOrder(createdByOrder)
-			.IsRequired();
-
-		builder.Property(e => e.EditedBy)
-			.HasColumnOrder(editedByOrder)
-			.IsRequired(false);
+			.IsClustered(clustered);
 	}
 
 	/// <summary>
