+++ src/BB84.EntityFrameworkCore.Repositories.SqlServer/Configurations/EnumeratorConfiguration.cs
@@ -1,4 +1,4 @@
-﻿// Copyright: 2024 Robert Peter Meyer
+// Copyright: 2024 Robert Peter Meyer
 // License: MIT
 //
 // This source code is licensed under the MIT license found in the
@@ -10,75 +10,34 @@ using BB84.EntityFrameworkCore.Entities.Abstractions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
+using Base = BB84.EntityFrameworkCore.Repositories.Configurations;
+
 namespace BB84.EntityFrameworkCore.Repositories.SqlServer.Configurations;
 
-/// <summary>
-/// Represents an abstract base class for configuring entity types that implement the
-/// <see cref="IEnumeratorEntity{Tkey}"/> interface.
-/// </summary>
+/// <inheritdoc cref="Base.EnumeratorConfiguration{TEntity, TKey}"/>
 /// <remarks>
-/// <para>
-/// This class defines a standard configuration for entities, including primary key setup,
-/// concurrency tokens, property constraints and indexing.
-/// </para>
-/// <para>
-/// A global query filter excluding soft deleted rows is applied, so that entities marked as
-/// deleted by the soft deletable interceptor are no longer returned by queries. Pass
-/// <see langword="true"/> for the <c>ignoreQueryFilters</c> parameter of the repository read
-/// methods to include them again.
-/// </para>
-/// <para>
-/// Be aware that a <b>required</b> navigation pointing at a soft deletable entity type
-/// interacts badly with this filter: filtering out the principal row also removes the
-/// dependents that require it. Model such navigations as optional, or apply a matching filter
-/// to both ends of the relationship.
-/// </para>
+/// Applies the provider-agnostic configuration of
+/// <see cref="Base.EnumeratorConfiguration{TEntity, TKey}"/> and tunes it for SQL Server by
+/// declaring the primary key as non clustered.
 /// </remarks>
-/// <typeparam name="TEntity">The type of the entity being configured.</typeparam>
-/// <typeparam name="TKey">The type of the key for the entity.</typeparam>
 [SuppressMessage("Style", "IDE0058", Justification = "Not relevant here, entity type configuration.")]
-public abstract class EnumeratorConfiguration<TEntity, TKey> : IEntityTypeConfiguration<TEntity>
+public abstract class EnumeratorConfiguration<TEntity, TKey> : Base.EnumeratorConfiguration<TEntity, TKey>, IEntityTypeConfiguration<TEntity>
 	where TEntity : class, IEnumeratorEntity<TKey>
 	where TKey : IEquatable<TKey>
 {
 	/// <inheritdoc/>
-	public virtual void Configure(EntityTypeBuilder<TEntity> builder)
+	public override void Configure(EntityTypeBuilder<TEntity> builder)
 	{
-		builder.HasKey(x => x.Id)
-			.IsClustered(false);
+		base.Configure(builder);
 
-		builder.Property(e => e.Id)
-			.HasColumnOrder(1)
-			.IsRequired();
-
-		EntityTypeBuilderDefaults.ApplyConcurrencyToken(builder, columnOrder: 2);
-
-		builder.Property(e => e.Name)
-			.HasColumnOrder(3)
-			.HasMaxLength(64)
-			.IsRequired()
-			.IsUnicode(false);
-
-		builder.Property(e => e.Description)
-			.HasColumnOrder(4)
-			.HasMaxLength(256)
-			.IsRequired(false)
-			.IsUnicode();
-
-		builder.Property(e => e.IsDeleted)
-			.HasColumnOrder(5)
-			.HasDefaultValue(false);
-
-		builder.HasQueryFilter(e => !e.IsDeleted);
-
-		builder.HasIndex(e => e.Name)
-			.IsUnique();
+		EntityTypeBuilderDefaults.ApplyKeyClustering<TEntity, TKey>(builder, clustered: false);
 	}
 }
 
 /// <inheritdoc cref="EnumeratorConfiguration{TEntity, TKey}"/>
 /// <remarks>
-/// The identity column is of type <see cref="int"/>.
+/// The identity column is of type <see cref="int"/> and the primary key is clustered, which
+/// suits the narrow, densely packed lookup tables this rung is meant for.
 /// </remarks>
 [SuppressMessage("Style", "IDE0058", Justification = "Not relevant here, entity type configuration.")]
 public abstract class EnumeratorConfiguration<TEntity> : EnumeratorConfiguration<TEntity, int>
@@ -89,7 +48,6 @@ public abstract class EnumeratorConfiguration<TEntity> : EnumeratorConfiguration
 	{
 		base.Configure(builder);
 
-		builder.HasKey(x => x.Id)
-			.IsClustered();
+		EntityTypeBuilderDefaults.ApplyKeyClustering<TEntity, int>(builder, clustered: true);
 	}
 }
