+++ src/BB84.EntityFrameworkCore.Repositories.SqlServer/Configurations/IdentityConfiguration.cs
@@ -1,4 +1,4 @@
-﻿// Copyright: 2024 Robert Peter Meyer
+// Copyright: 2024 Robert Peter Meyer
 // License: MIT
 //
 // This source code is licensed under the MIT license found in the
@@ -10,29 +10,27 @@ using BB84.EntityFrameworkCore.Entities.Abstractions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
+using Base = BB84.EntityFrameworkCore.Repositories.Configurations;
+
 namespace BB84.EntityFrameworkCore.Repositories.SqlServer.Configurations;
 
-/// <summary>
-/// Represents an abstract base class for configuring entity types that implement the
-/// <see cref="IIdentityEntity{Tkey}"/> interface.
-/// </summary>
+/// <inheritdoc cref="Base.IdentityConfiguration{TEntity, TKey}"/>
 /// <remarks>
-/// This class defines a default configuration for identity-related entities, including the primary key and
-/// concurrency token. Derived classes can override the <see cref="Configure"/> method to provide additional
-/// or customized configuration.
+/// Applies the provider-agnostic configuration of
+/// <see cref="Base.IdentityConfiguration{TEntity, TKey}"/> and tunes it for SQL Server by
+/// declaring the primary key as non clustered.
 /// </remarks>
-/// <typeparam name="TEntity">The type of the entity being configured.</typeparam>
-/// <typeparam name="TKey">The type of the primary key for the entity.</typeparam>
 [SuppressMessage("Style", "IDE0058", Justification = "Not relevant here, entity type configuration.")]
-public abstract class IdentityConfiguration<TEntity, TKey> : IEntityTypeConfiguration<TEntity>
+public abstract class IdentityConfiguration<TEntity, TKey> : Base.IdentityConfiguration<TEntity, TKey>, IEntityTypeConfiguration<TEntity>
 	where TEntity : class, IIdentityEntity<TKey>
 	where TKey : IEquatable<TKey>
 {
 	/// <inheritdoc/>
-	public virtual void Configure(EntityTypeBuilder<TEntity> builder)
+	public override void Configure(EntityTypeBuilder<TEntity> builder)
 	{
-		EntityTypeBuilderDefaults.ApplyIdentityKey<TEntity, TKey>(builder);
-		EntityTypeBuilderDefaults.ApplyConcurrencyToken(builder, columnOrder: 2);
+		base.Configure(builder);
+
+		EntityTypeBuilderDefaults.ApplyKeyClustering<TEntity, TKey>(builder, clustered: false);
 	}
 }
