+++ src/BB84.EntityFrameworkCore.Repositories.SqlServer/README.md
@@ -3,12 +3,33 @@
 
 # BB84.EntityFrameworkCore.Repositories.SqlServer
 
-This package provides SQL Server–specific entity type configuration base classes, `SaveChangesInterceptor` implementations for auditing and soft delete, and `EntityTypeBuilderExtensions` for temporal table support.
+This package provides the SQL Server tuning of the entity type configuration base classes, `PropertyBuilderExtensions` for column types, `EntityTypeBuilderExtensions` for temporal table support, and `DatabaseFacadeExtensions` for calling stored procedures and SQL functions.
+
+> **Moved in 5.0.** The configuration base classes now derive from provider-agnostic bases in
+> `BB84.EntityFrameworkCore.Repositories`, and the interceptors moved there outright. `DatabaseFacadeExtensions` moved
+> the other way, into this package, because it emits T-SQL. See [Package boundaries](#package-boundaries).
 
 ## Installation
 
 ```powershell
 dotnet add package BB84.EntityFrameworkCore.Repositories.SqlServer
+```
+
+## Package boundaries
+
+Each configuration ladder exists twice, under the **same type names**:
+
+| Namespace                                                | Applies                                                                     |
+| -------------------------------------------------------- | --------------------------------------------------------------------------- |
+| `BB84.EntityFrameworkCore.Repositories.Configurations`   | Key declaration, column order, concurrency token, audit columns, soft delete filter |
+| `BB84.EntityFrameworkCore.Repositories.SqlServer.Configurations` | The above, plus clustering, `NEWID()` defaults and `sysname` audit columns |
+
+Pick the namespace, not the type. On SQL Server, keep importing this package's namespace and nothing changes. On any other provider, import the agnostic one and you get everything except the three SQL Server calls.
+
+A file that imports **both** namespaces — a project mapping a SQL Server context and a second provider's context side by side — hits `CS0104` on the ambiguous name. Alias one of them:
+
+```csharp
+using SqlServerConfigurations = BB84.EntityFrameworkCore.Repositories.SqlServer.Configurations;
 ```
 
 ## Configuration base classes
@@ -132,31 +153,54 @@ builder.ToHistoryTable(
 
 ## Interceptors
 
-Register interceptors on the `DbContextOptionsBuilder`:
+`SoftDeletableInterceptor` and `TimeAuditedInterceptor` moved to `BB84.EntityFrameworkCore.Repositories` in 5.0 — nothing in them was SQL Server specific. Update the `using` to `BB84.EntityFrameworkCore.Repositories.Interceptors`; the types and their behavior are unchanged.
+
+## `DatabaseFacadeExtensions`
+
+Extension methods on `DatabaseFacade` (`context.Database`) for calling SQL Server stored procedures and functions safely with parameterized SQL. These moved here from `BB84.EntityFrameworkCore.Repositories` in 5.0, because the SQL they emit — `EXECUTE [schema].[name] @p = @p OUTPUT`, `SELECT [Value] = [schema].[fn](@p)` — is T-SQL by construction.
+
+### Stored procedures
 
 ```csharp
-services.AddSingleton<SoftDeletableInterceptor>();
-services.AddSingleton<TimeAuditedInterceptor>();
+// Returns rows as IReadOnlyList<T>
+IReadOnlyList<ReportRow> rows = context.Database.ExecuteProcedure<ReportRow>(
+    schema: "dbo",
+    name: "usp_GetReport",
+    parameters: [new SqlParameter("@FromDate", fromDate)]);
 
-services.AddDbContext<AppDbContext>((sp, options) =>
-{
-    options
-        .UseSqlServer(connectionString)
-        .AddInterceptors(
-            sp.GetRequiredService<SoftDeletableInterceptor>(),
-            sp.GetRequiredService<TimeAuditedInterceptor>());
-});
+// Non-generic overload returns rows-affected count
+int affected = context.Database.ExecuteProcedure(
+    schema: "dbo",
+    name: "usp_ArchiveOrders",
+    parameters: [new SqlParameter("@CutoffDate", cutoff)]);
+
+// Async variants
+await context.Database.ExecuteProcedureAsync<ReportRow>(..., cancellationToken);
+await context.Database.ExecuteProcedureAsync(..., cancellationToken);
 ```
 
-### `SoftDeletableInterceptor`
+Output parameters are supported — parameters with `Direction = ParameterDirection.Output` are emitted as `@param = @param OUTPUT` in the generated SQL.
 
-Fires on `SavingChanges`/`SavingChangesAsync`. For every entity tracked as `Deleted` that implements `ISoftDeletable`, it sets `IsDeleted = true` and changes the state to `Modified` — preventing a physical `DELETE` from being issued.
+### Table-valued functions
 
-### `TimeAuditedInterceptor`
+```csharp
+IReadOnlyList<ProductDto> results = context.Database.ExecuteTableFunction<ProductDto>(
+    schema: "dbo",
+    name: "fn_GetActiveProducts",
+    parameters: [new SqlParameter("@CategoryId", categoryId)]);
 
-Fires on `SavingChanges`/`SavingChangesAsync`. For every entity implementing `ITimeAudited`:
+await context.Database.ExecuteTableFunctionAsync<ProductDto>(..., cancellationToken);
+```
 
-- `EntityState.Added` → sets `CreatedAt = DateTimeOffset.UtcNow`
-- `EntityState.Modified` → sets `EditedAt = DateTimeOffset.UtcNow`
+### Scalar-valued functions
 
-`CreatedBy` / `EditedBy` are **not** set automatically by this interceptor — implement a custom `SaveChangesInterceptor` for user auditing and register it alongside the built-in ones.
+```csharp
+decimal? total = context.Database.ExecuteScalarFunction<decimal>(
+    schema: "dbo",
+    name: "fn_GetOrderTotal",
+    parameters: [new SqlParameter("@OrderId", orderId)]);
+
+await context.Database.ExecuteScalarFunctionAsync<decimal>(..., cancellationToken);
+```
+
+All methods sanitize schema/name inputs and use parameterized SQL to prevent injection. The parameter sequence is materialized once on entry, so a lazily generated `IEnumerable<DbParameter>` is safe to pass.
