+++ tests/BB84.EntityFrameworkCore.Repositories.Tests/Persistence/TestDbContext.cs
@@ -3,7 +3,7 @@
 //
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
-using BB84.EntityFrameworkCore.Repositories.SqlServer.Interceptors;
+using BB84.EntityFrameworkCore.Repositories.Interceptors;
 using BB84.EntityFrameworkCore.Repositories.Tests.Abstractions;
 using BB84.EntityFrameworkCore.Repositories.Tests.Persistence.Interceptors;
